+++ modules/sagemaker/sagemaker-studio/stack.py
@@ -34,6 +34,8 @@ class SagemakerStudioStack(Stack):
         enable_domain_resource_isolation: bool,
         enable_jupyterlab_app: bool,
         enable_jupyterlab_app_sharing: bool,
+        enable_docker_access: bool,
+        vpc_only_trusted_accounts: List[str],
         jupyterlab_app_instance_type: Optional[str],
         auth_mode: str,
         role_path: Optional[str],
@@ -75,6 +77,8 @@ class SagemakerStudioStack(Stack):
             app_image_config_name=app_image_config_name,
             image_name=image_name,
             auth_mode=auth_mode,
+            enable_docker_access=enable_docker_access,
+            vpc_only_trusted_accounts=vpc_only_trusted_accounts,
         )
 
         if enable_custom_sagemaker_projects:
@@ -326,6 +330,8 @@ class SagemakerStudioStack(Stack):
         app_image_config_name: Optional[str],
         image_name: Optional[str],
         auth_mode: str,
+        enable_docker_access: bool,
+        vpc_only_trusted_accounts: List[str],
     ) -> sagemaker.CfnDomain:
         """
         Create the SageMaker Studio Domain
@@ -334,7 +340,12 @@ class SagemakerStudioStack(Stack):
         :param s3_bucket: - S3 bucket used for sharing notebooks between users
         :param sagemaker_studio_role: - IAM Execution Role for the domain
         :param subnet_ids: - list of comma separated subnet ids
-        :param vpc_id: - VPC Id for the domain
+        :param vpc_id: -  VPC Id for the domain
+        :param app_image_config_name: - config name
+        :param image_name: - image name
+        :param auth_mode: - auth mode (IAM or SSO)
+        :param enable_docker_access: - flag to enable docker from studio
+        :param vpc_only_trusted_accounts: - list of trusted aws account ids in vpc only mode
         """
         sagemaker_sg = ec2.SecurityGroup(
             self,
@@ -367,6 +378,12 @@ class SagemakerStudioStack(Stack):
                 security_groups=[sagemaker_sg.security_group_id],
                 sharing_settings=sagemaker.CfnDomain.SharingSettingsProperty(),
                 **custom_kernel_settings,  # type:ignore
+            ),
+            domain_settings=sagemaker.CfnDomain.DomainSettingsProperty(
+                docker_settings=sagemaker.CfnDomain.DockerSettingsProperty(
+                    enable_docker_access="ENABLED" if enable_docker_access else "DISABLED",
+                    vpc_only_trusted_accounts=vpc_only_trusted_accounts if vpc_only_trusted_accounts else None,
+                ),
             ),
             domain_name=domain_name,
             subnet_ids=subnet_ids,