+++ modules/sagemaker/sagemaker-studio/tests/test_stack.py
@@ -49,6 +49,8 @@ def stack(
     mlflow_artifact_store_bucket_name = "bucket"
     mlflow_artifact_store_bucket_prefix = "/"
     enable_jupyterlab_app_sharing = False
+    enable_docker_access = False
+    vpc_only_trusted_accounts = []
     jupyterlab_app_instance_type = None
 
     return stack.SagemakerStudioStack(
@@ -65,6 +67,8 @@ def stack(
         enable_custom_sagemaker_projects=enable_custom_sagemaker_projects,
         enable_jupyterlab_app=enable_jupyterlab_app,
         enable_jupyterlab_app_sharing=enable_jupyterlab_app_sharing,
+        enable_docker_access=enable_docker_access,
+        vpc_only_trusted_accounts=vpc_only_trusted_accounts,
         jupyterlab_app_instance_type=jupyterlab_app_instance_type,
         auth_mode=auth_mode,
         role_path=role_path,